fix(determinism-check): split test module out of determinism_check.rs
sentrux god-file gate failed self-scan: functions>25 && loc>400
(26 fns, 672 loc). Move #[cfg(test)] mod tests to a separate file via
#[path], matching the existing convention across the codebase.
